+++ packages/main/src/components/AnalyticalTable/TableBody/VirtualTableBody.tsx
@@ -3,8 +3,8 @@ import { useVirtualizer } from '@tanstack/react-virtual';
 import { clsx } from 'clsx';
 import type { MutableRefObject, ReactNode } from 'react';
 import React, { useCallback, useMemo, useRef } from 'react';
-import type { AnalyticalTablePropTypes, DivWithCustomScrollProp } from '../index.js';
 import type { ScrollToRefType } from '../interfaces.js';
+import type { AnalyticalTablePropTypes, DivWithCustomScrollProp } from '../types/index.js';
 import { getSubRowsByString } from '../util/index.js';
 import { EmptyRow } from './EmptyRow.js';
 import { RowSubComponent as SubComponent } from './RowSubComponent.js';
