--- packages/main/src/components/AnalyticalTable/types/FilterEntry.ts
@@ -1,4 +1,0 @@
-export interface FilterEntry {
-  id: string;
-  value: string | number;
-}
